pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "globalmart-fashion-assistant"
version = "0.1.0"
requires-python = ">=3.10"
description = "GlobalMart Fashion multimodal assistant demo (Cohere + RAG)"
readme = "README.md"
dependencies = [
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "python-multipart>=0.0.9",
    "faster-whisper>=1.1.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
